requirements.txt
pandas
requests
minio==7.1.14
apache-airflow-providers-docker==3.7.1
apache-airflow-providers-slack==8.1.0
apache-airflow-providers-microsoft-azure==8.0.0
azure-mgmt-datafactory==1.0.0
psycopg2-binary
pymysql
sqlalchemy
selenium==4.18.1
boto3==1.34.34
botocore==1.34.34